cli/requirements/requirements-docs.txt
#
# This file is autogenerated by hatch-pip-compile with Python 3.11
#
# [constraints] requirements.txt (SHA256: 2a92e1c6e109ef4c9ce314b9a60b303169d2ddf6cb3b4c4c9df4288cec4ac2ef)
#
# - markdown-callouts
# - markdown-exec
# - mkdocs
# - mkdocs-click
# - mkdocs-exclude-search
# - mkdocs-gen-files
# - mkdocs-literate-nav
# - mkdocs-material
# - mkdocs-section-index
# - mkdocstrings
# - mkdocstrings-python
# - pymdown-extensions
# - click~=8.1.3
# - fake-useragent~=1.4.0
# - pandas<3,>=2
# - pydantic~=1.10.13
# - python-dotenv~=1.0.0
# - pytz~=2023.2
# - pyyaml~=6.0
# - ratelimit~=2.2.1
# - requests~=2.31.0
# - rich-click~=1.6.1
# - rich~=13.3.2
# - tenacity~=8.2.2
#

babel==2.14.0
    # via mkdocs-material
certifi==2023.11.17
    # via
    #   -c requirements.txt
    #   requests
charset-normalizer==3.3.2
    # via
    #   -c requirements.txt
    #   requests
click==8.1.7
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
    #   mkdocs
    #   mkdocs-click
    #   mkdocstrings
    #   rich-click
colorama==0.4.6
    # via
    #   griffe
    #   mkdocs-material
fake-useragent==1.4.0
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
ghp-import==2.1.0
    # via mkdocs
griffe==0.38.1
    # via mkdocstrings-python
idna==3.6
    # via
    #   -c requirements.txt
    #   requests
jinja2==3.1.2
    # via
    #   mkdocs
    #   mkdocs-material
    #   mkdocstrings
markdown==3.5.1
    # via
    #   -c requirements.txt
    #   markdown-callouts
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocs-click
    #   mkdocs-material
    #   mkdocstrings
    #   pymdown-extensions
markdown-callouts==0.3.0
    # via hatch.envs.docs
markdown-exec==1.8.0
    # via hatch.envs.docs
markdown-it-py==2.2.0
    # via
    #   -c requirements.txt
    #   rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   mkdocs
    #   mkdocstrings
mdurl==0.1.2
    # via
    #   -c requirements.txt
    #   markdown-it-py
mergedeep==1.3.4
    # via mkdocs
mkdocs==1.5.3
    # via
    #   hatch.envs.docs
    #   mkdocs-autorefs
    #   mkdocs-exclude-search
    #   mkdocs-gen-files
    #   mkdocs-literate-nav
    #   mkdocs-material
    #   mkdocs-section-index
    #   mkdocstrings
mkdocs-autorefs==0.5.0
    # via mkdocstrings
mkdocs-click==0.8.1
    # via hatch.envs.docs
mkdocs-exclude-search==0.6.6
    # via hatch.envs.docs
mkdocs-gen-files==0.5.0
    # via hatch.envs.docs
mkdocs-literate-nav==0.6.1
    # via hatch.envs.docs
mkdocs-material==9.5.3
    # via hatch.envs.docs
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mkdocs-section-index==0.3.8
    # via hatch.envs.docs
mkdocstrings==0.24.0
    # via
    #   hatch.envs.docs
    #   mkdocstrings-python
mkdocstrings-python==1.7.5
    # via hatch.envs.docs
numpy==1.26.3
    # via
    #   -c requirements.txt
    #   pandas
packaging==23.2
    # via mkdocs
paginate==0.5.6
    # via mkdocs-material
pandas==2.1.4
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
pathspec==0.12.1
    # via mkdocs
platformdirs==4.1.0
    # via
    #   mkdocs
    #   mkdocstrings
pydantic==1.10.13
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
pygments==2.17.2
    # via
    #   -c requirements.txt
    #   mkdocs-material
    #   rich
pymdown-extensions==10.7
    # via
    #   hatch.envs.docs
    #   markdown-exec
    #   mkdocs-material
    #   mkdocstrings
python-dateutil==2.8.2
    # via
    #   -c requirements.txt
    #   ghp-import
    #   pandas
python-dotenv==1.0.0
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
pytz==2023.3.post1
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
    #   pandas
pyyaml==6.0.1
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
    #   mkdocs
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
ratelimit==2.2.1
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
regex==2023.12.25
    # via mkdocs-material
requests==2.31.0
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
    #   mkdocs-material
rich==13.3.5
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
    #   rich-click
rich-click==1.6.1
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
six==1.16.0
    # via
    #   -c requirements.txt
    #   python-dateutil
tenacity==8.2.3
    # via
    #   -c requirements.txt
    #   hatch.envs.docs
typing-extensions==4.9.0
    # via
    #   -c requirements.txt
    #   pydantic
tzdata==2023.4
    # via
    #   -c requirements.txt
    #   pandas
urllib3==2.1.0
    # via
    #   -c requirements.txt
    #   requests
watchdog==3.0.0
    # via mkdocs
